% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{add_R}
\alias{add_R}
\title{Append R to string, if it doesn't end in R already.}
\usage{
add_R(x)
}
\arguments{
\item{x}{a string}
}
\description{
Use this function to conveniently rename reverse-coded variables, so that
they end in R.
}
\examples{
data('bfi')
bfi \%>\% dplyr::select(BFIK_open_2,BFIK_agree_2) \%>\% dplyr::rename_at(1, add_R) \%>\% head()
}
