% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bfi}
\alias{bfi}
\title{mock BFI data}
\format{A data frame with 28 rows and 29 variables:
\describe{
\item{session}{formr user code}
\item{created}{Created date time}
\item{BFIK_agree_4R}{An agreeableness item}
}}
\usage{
bfi
}
\description{
a small mock BFI dataset with realistic values, exported from formr.
The dataset is self-documenting via attributes.
}
\keyword{datasets}
