% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_component_fallback}
\alias{codebook_component_fallback}
\title{codebook component for scales}
\usage{
codebook_component_fallback(item, results, item_name, indent = "###")
}
\arguments{
\item{item}{an item without attributes set}

\item{results}{a formr results table with attributes set on items and scales}

\item{item_name}{the item name}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
codebook component for scales
}
