% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_items}
\alias{codebook_items}
\title{codebook items}
\usage{
codebook_items(results, indent = "##")
}
\arguments{
\item{results}{a formr results table which has the following columns: session, created, modified, expired, ended}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
you need the formr package to use this function, and the item data has to be encoded in the item attribute
}
