% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_jsonld}
\alias{codebook_jsonld}
\title{codebook missingness}
\usage{
codebook_jsonld(results)
}
\arguments{
\item{results}{a formr results table which has the following columns: session, created, modified, expired, ended}
}
\description{
codebook missingness
}
