% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_missingness}
\alias{codebook_missingness}
\title{Codebook missingness}
\usage{
codebook_missingness(results, indent = "##")
}
\arguments{
\item{results}{a formr results table which has the following columns: session, created, modified, expired, ended}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
Codebook missingness
}
