% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_survey_overview}
\alias{codebook_survey_overview}
\title{Codebook survey overview}
\usage{
codebook_survey_overview(results, survey_repetition = "single",
  indent = "##")
}
\arguments{
\item{results}{a formr results table which has the following columns: session, created, modified, expired, ended}

\item{survey_repetition}{defaults to single (other values: repeated_once, repeated_many). controls whether internal consistency, retest reliability or multilevel reliability is computed}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
Codebook survey overview
}
