% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_table}
\alias{codebook_table}
\title{codebook metadata table}
\usage{
codebook_table(results)
}
\arguments{
\item{results}{a data frame}
}
\description{
will generate a table combining metadata from variable attributes
with data summaries generated using skimr
}
