% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{compute_reliabilites}
\alias{compute_reliabilites}
\title{Compute reliabilities}
\usage{
compute_reliabilites(results, survey_repetition = "single")
}
\arguments{
\item{results}{a formr results table with attributes set on items and scales}

\item{survey_repetition}{defaults to "single". Can also be "repeated_once" or "repeated_many"}
}
\description{
If you pass the object resulting from a call to formr_results to this function, it will compute reliabilities for each scale.
Internally, each reliability computation is passed to a future. If you are calculating multilevel reliabilities, it may be worthwhile to parallelise this operation using future::plan
}
\examples{
# see vignette
}
