% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_citation<-}
\alias{data_citation<-}
\title{Add citation information to a dataset}
\usage{
data_citation(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the citation information}
}
\description{
Add a single character vector or a list to give citation information for
JSON-LD metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data('bfi')
data_citation(bfi) <- "Arslan (2018). Mock BFI data."
attributes(bfi)[[4]]

}
