% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_creator<-}
\alias{data_creator<-}
\title{Add creator}
\usage{
data_creator(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the keywords}
}
\description{
Add a character vector to to add keywords in preparation for JSON-LD
metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data('bfi')
data_creator(bfi) <- Sys.time()
attributes(bfi)[[4]]

}
