% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_description<-}
\alias{data_description<-}
\title{Add description to a dataset}
\usage{
data_description(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the description}
}
\description{
Add a single character vector to describe a data frame in preparation for JSON-LD
metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data_description(bfi) <- "a small mock Big Five Inventory dataset"

}
