% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_description_default}
\alias{data_description_default}
\title{Data description default}
\usage{
data_description_default(data)
}
\arguments{
\item{data}{the data frame}
}
\description{
If you do not define a dataset description yourself, this will be the
automatically generated default.
}
\examples{
data('bfi')
data_description_default(bfi)

}
