% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_identifier<-}
\alias{data_identifier<-}
\title{Add identifier to a dataset}
\usage{
data_identifier(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the identifier}
}
\description{
Add a single character vector or a list to give an identifier for the dataset,
such as a DOI.
}
\examples{
data_identifier(bfi) <- "doi:10.5281/zenodo.1326520"

}
