% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_keywords<-}
\alias{data_keywords<-}
\title{Add keywords to a dataset}
\usage{
data_keywords(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the keywords}
}
\description{
Add a character vector to to add keywords in preparation for JSON-LD
metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data('bfi')
data_keywords(bfi) <- c("Big Five", "Personality", "Psychology")
attributes(bfi)[[4]]

}
