% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_name<-}
\alias{data_name<-}
\title{Give a name to a dataset}
\usage{
data_name(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the name}
}
\description{
Add a single character vector to name a data frame in preparation for JSON-LD
metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data('bfi')
data_name(bfi) <- "a small mock Big Five Inventory dataset"
attributes(bfi)[[4]]

}
