% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_spatialCoverage<-}
\alias{data_spatialCoverage<-}
\title{Add spatial coverage information to a dataset}
\usage{
data_spatialCoverage(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the spatial coverage information}
}
\description{
Add a single character vector or a list to give spatial coverage the dataset
if it has a spatial extent for JSON-LD metadata generation, e.g., the place
where data was collected.
}
\examples{
data('bfi')
data_spatialCoverage(bfi) <- "Goettingen, Germany"
attributes(bfi)[[4]]

}
