% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_temporalCoverage<-}
\alias{data_temporalCoverage<-}
\title{Add temporal coverage information to a dataset}
\usage{
data_temporalCoverage(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the temporal coverage information}
}
\description{
Add a single character vector or a list to give temporal coverage the dataset
if it has a temporal extent for JSON-LD metadata generation, e.g., the period
of data collection.
}
\examples{
data_temporalCoverage(bfi) <- "2017"

}
