% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{data_url<-}
\alias{data_url<-}
\title{Add an URL to a dataset}
\usage{
data_url(data) <- value
}
\arguments{
\item{data}{the data frame}

\item{value}{the URL}
}
\description{
Add a URL where the dataset description can be found in preparation for
JSON-LD metadata generation using \code{\link[=codebook]{codebook()}} or \code{\link[=metadata_list]{metadata_list()}}.
}
\examples{
data('bfi')
data_url(bfi) <- "https://rubenarslan.github.io/codebook/articles/codebook.html"
attributes(bfi)[[4]]

}
