% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ended}
\alias{ended}
\title{How many surveys were ended?}
\usage{
ended(survey, variable = "ended")
}
\arguments{
\item{survey}{which survey are you asking about?}

\item{variable}{which variable should be filled out, defaults to "ended"}
}
\description{
Just a simple to check how many times a survey (e.g. diary)
was finished. It defaults to checking the "ended" variable for this.
}
\examples{
survey <- data.frame(ended = c("2016-05-28 10:11:00", NA, "2016-05-30 11:18:28"))
ended(survey = survey)
}
