% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_table.R
\name{get_skimmers.haven_labelled_spss}
\alias{get_skimmers.haven_labelled_spss}
\title{Define skimmers for haven_labelled_spss variables}
\usage{
get_skimmers.haven_labelled_spss(column)
}
\arguments{
\item{column}{the column to skim}
}
\description{
Variables labelled using the haven_labelled_spss class are special
because the underlying data can be numeric or character. This skimmers summarises
both and leaves non-pertinent columns missings.
}
