% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{knit_print.alpha}
\alias{knit_print.alpha}
\title{Generate rmarkdown codebook}
\usage{
knit_print.alpha(x, indent = "###")
}
\arguments{
\item{x}{a psych alpha object}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h4}
}
\description{
If you pass the object resulting from a call to formr_results to this function, it will generate a markdown codebook for this object.
}
\examples{
example("alpha", "psych")
knit_print.alpha(a4)

}
