% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{knit_print.multilevel}
\alias{knit_print.multilevel}
\title{Print a psych::multilevel.reliability object for knitr}
\usage{
knit_print.multilevel(x, indent = "#####", ...)
}
\arguments{
\item{x}{a psych alpha object}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h5}

\item{...}{ignored}
}
\description{
Just prints the normal output.
}
\examples{
example("mlr", "psych")
knitr::knit_print(m4)

}
