% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{knit_print.scaleDiagnosis}
\alias{knit_print.scaleDiagnosis}
\title{Print a \code{\link[userfriendlyscience:scaleDiagnosis]{userfriendlyscience::scaleDiagnosis()}} object for knitr}
\usage{
\method{knit_print}{scaleDiagnosis}(x, indent = "#####", ...)
}
\arguments{
\item{x}{a scaleDiagnosis object}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h5}

\item{...}{ignored}
}
\description{
Just prints the normal output of \code{\link[userfriendlyscience:scaleDiagnosis]{userfriendlyscience::scaleDiagnosis()}}.
}
\examples{
example("mlr", "psych")
knitr::knit_print(mg)

}
