% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_browser.R
\name{label_browser}
\alias{label_browser}
\title{Browse and search variable and value labels}
\usage{
label_browser()
}
\description{
Same as the \code{\link[=codebook_browser]{codebook_browser()}}, but doesn't show data summaries and
additional attributes.
}
