% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_browser.R
\name{label_browser_static}
\alias{label_browser_static}
\title{Browse and search variable and value labels}
\usage{
label_browser_static(data = NULL, viewer = rstudioapi::viewer)
}
\arguments{
\item{data}{data frame. if left empty, will use the text you currently select in RStudio as the label or the first data frame in your environment}

\item{viewer}{where to show. defaults to viewer tab}
}
\description{
Same as the \code{\link[=codebook_browser]{codebook_browser()}}, but doesn't show data summaries and
additional attributes. This yields a static table, so you can continue
to edit code while viewing the labels, but you cannot switch the dataset
via a dropdown menu.
}
\examples{
label_browser_static(bfi)


}
