% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{load_data_and_render_codebook}
\alias{load_data_and_render_codebook}
\title{render codebook based on file}
\usage{
load_data_and_render_codebook(file, text, ...)
}
\arguments{
\item{file}{file to make codebook from (sav, rds, dta, etc.)}

\item{text}{codebook template}

\item{...}{all other arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Render text
}
