% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{metadata_jsonld}
\alias{metadata_jsonld}
\title{Metadata as JSON-LD}
\usage{
metadata_jsonld(results)
}
\arguments{
\item{results}{a formr results table which has the following columns: session, created, modified, expired, ended}
}
\description{
Metadata as JSON-LD
}
