% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{metadata_list}
\alias{metadata_list}
\title{Metadata from dataframe}
\usage{
metadata_list(results, name = NULL, only_existing = TRUE)
}
\arguments{
\item{results}{a data frame, ideally with attributes set on variables}

\item{name}{the name of the data frame (taken from attributes if set or from
name of first argument)}

\item{only_existing}{whether to drop helpful metadata to comply with the list
of currently defined schema.org properties}
}
\description{
Returns a list containing variable metadata (attributes) and data summaries.
}
\examples{
data("bfi")
md_list <- metadata_list(bfi)
md_list$variableMeasured[[20]]
}
