% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{new_codebook_rmd}
\alias{new_codebook_rmd}
\title{Create a codebook rmarkdown document}
\usage{
new_codebook_rmd(filename = "codebook", template = "default")
}
\arguments{
\item{filename}{under which file name do you want to create a template}

\item{template}{only "default" exists for now}
}
\description{
This function will create and open an .Rmd file in the current working
directory. By default, the file is named codebook.Rmd. No files will be
overwritten. The .Rmd file has some useful defaults set for creating codebooks.
}
\examples{
\dontrun{
new_codebook_rmd()
}
}
