% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{plot_labelled}
\alias{plot_labelled}
\title{Plot labelled vector}
\usage{
plot_labelled(item, item_name = deparse(substitute(item)), wrap_at = 50,
  go_vertical = FALSE)
}
\arguments{
\item{item}{a vector}

\item{item_name}{item name, defaults to name of first argument}

\item{wrap_at}{the subtitle (the label) will be wrapped at this number of characters}

\item{go_vertical}{defaults to FALSE. Whether to show choices on the Y axis instead.}
}
\description{
Plot a labelled vector, making use of the variable name, label and value
labels to make the plot more readable. This function also works for other
vectors, but provides little benefit.
}
\examples{
data("bfi", package = "codebook")
plot_labelled(bfi$BFIK_open_1)
}
