% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{print.knit_asis}
\alias{print.knit_asis}
\title{Print new lines in \code{knit_asis} outputs}
\usage{
\method{print}{knit_asis}(x, ...)
}
\arguments{
\item{x}{the knit_asis object}

\item{...}{ignored}
}
\description{
Print new lines in \code{knit_asis} outputs
}
