% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_factor}
\alias{zap_label}
\alias{var_label}
\alias{var_label<-}
\alias{val_label}
\alias{val_label<-}
\alias{val_labels}
\alias{val_labels<-}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}}

  \item{forcats}{\code{\link[forcats]{as_factor}}}

  \item{haven}{\code{\link[haven]{zap_label}}}

  \item{labelled}{\code{\link[labelled:val_labels]{val_label}}, \code{\link[labelled:val_labels]{val_label<-}}, \code{\link[labelled]{val_labels}}, \code{\link[labelled:val_labels]{val_labels<-}}, \code{\link[labelled]{var_label}}, \code{\link[labelled:var_label]{var_label<-}}}
}}

