% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_attributes.R
\name{rescue_attributes}
\alias{rescue_attributes}
\title{Rescue lost attributes}
\usage{
rescue_attributes(df_no_attributes, df_with_attributes)
}
\arguments{
\item{df_no_attributes}{the data frame with missing attributes}

\item{df_with_attributes}{the data frame from which you want to restore attributes}
}
\description{
You can use this function if some of your items have lost their attributes during wrangling
Variables have to have the same name (Duh) and no attributes should be overwritten.
But use with care. Similar to \code{\link[labelled:copy_labels]{labelled::copy_labels()}}.
}
