% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook_table.R
\name{skim_codebook}
\alias{skim_codebook}
\title{Skim codebook}
\usage{
skim_codebook(data, ...)
}
\arguments{
\item{data}{the dataset to skim}

\item{...}{passed to \code{\link[skimr:skim]{skimr::skim()}}}
}
\description{
Implements the regular functionality of \code{\link[skimr:skim]{skimr::skim()}} but renames the columns
p0, p50, and p100 to min, median, and max respectively for numeric types to
keep things consistent across type (and produce a narrower wide table).
}
\examples{
skim_codebook(bfi)
}
