% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{summary.labelled}
\alias{summary.labelled}
\title{summary.labelled}
\usage{
\method{summary}{labelled}(object, ...)
}
\arguments{
\item{object}{a labelled vector}

\item{...}{passed to summary.factor}
}
\description{
summary.labelled
}
\examples{
example("labelled", "haven")
summary(x)

}
