% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\name{to_factor}
\alias{to_factor}
\title{To factor}
\usage{
to_factor(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{passed to \code{\link[haven:as_factor]{haven::as_factor()}}}
}
\description{
Convert a labelled vector to a factor, even if it doesn't have the proper
class, as long as it has the "labels" attribute.
You can have this attribute without, for example, the haven_labelled class,
when importing data with \code{\link[rio:import]{rio::import()}} for example.
}
\examples{
example("labelled", "haven")
to_factor(x)
to_factor(zap_labelled(x))
to_factor(as_factor(x))
to_factor(1:4)
}
