% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_attributes.R
\name{zap_attributes}
\alias{zap_attributes}
\title{zap attributes}
\usage{
zap_attributes(x, attributes = NULL)
}
\arguments{
\item{x}{the data frame or variable}

\item{attributes}{defaults to NULL. If character, only those attributes are zapped}
}
\description{
Modelled on havens zap_labels, but more flexible.
Can be used to zap all attributes for all variables in a
data frame, or for some variables, or some attributes.
}
\examples{
bfi <- data.frame(matrix(data = rnorm(300), ncol = 3))
names(bfi) <- c("bfi_e1", "bfi_e2R", "bfi_e3")
attributes(bfi$bfi_e1)$label <- "I am outgoing."
attributes(bfi$bfi_e2R)$label <- "I prefer books to people."
attributes(bfi$bfi_e3)$label <- "I love to party."
bfi$bfi_e <- rowMeans(bfi[, c("bfi_e1", "bfi_e2R", "bfi_e3")])
bfi <- detect_scales(bfi, quiet = TRUE) # create attributes
str(zap_attributes(bfi, "label"))
zap_attributes(bfi$bfi_e)
}
