% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_attributes.R
\name{zap_labelled}
\alias{zap_labelled}
\title{Zap labelled class}
\usage{
zap_labelled(x)
}
\arguments{
\item{x}{the data frame or variable}
}
\description{
Modelled on \code{\link[haven:zap_labels]{haven::zap_labels()}}, zaps labelled class (not other attributes).
}
