% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Generate rmarkdown codebook}
\usage{
codebook(results, reliabilities = NULL, survey_repetition = c("auto",
  "single", "repeated_once", "repeated_many"), missingness_report = TRUE,
  metadata_table = TRUE, metadata_json = TRUE, indent = "#")
}
\arguments{
\item{results}{a formr results table with attributes set on items and scales}

\item{reliabilities}{a named list with one entry per scale and one or several printable reliability computations for this scale. if NULL, computed on-the-fly using compute_reliabilities}

\item{survey_repetition}{defaults to "auto" which is to try to determine the level of repetition from the "session" and "created" variables. Other values are: single, repeated_once, repeated_many}

\item{missingness_report}{whether to print a missingness report. Turn off if this gets too complicated and you need a custom solution (e.g. in case of random missings).}

\item{metadata_table}{whether to print a metadata table/tabular codebook.}

\item{metadata_json}{whether to include machine-readable metadata as JSON-LD (not visible)}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
If you pass the object resulting from a call to formr_results to this function, it will generate a markdown codebook for this object.
}
\examples{
# see vignette
}
