% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{Generate rmarkdown codebook}
\usage{
codebook(results, indent = "#")
}
\arguments{
\item{results}{a formr results table with attributes set on items and scales}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
If you pass the object resulting from a call to formr_results to this function, it will generate a markdown codebook for this object.
}
\examples{
# see vignette
}
