% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_component_scale}
\alias{codebook_component_scale}
\title{codebook component for scales}
\usage{
codebook_component_scale(scale, scale_name, items, reliabilities,
  indent = "###")
}
\arguments{
\item{scale}{a scale with attributes set}

\item{scale_name}{the variable name of this scale}

\item{items}{a data.frame with the items constituting the scale}

\item{reliabilities}{a list with one or several results from calls to psych package functions for computing reliability}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
codebook component for scales
}
