% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_component_single_item}
\alias{codebook_component_single_item}
\title{Codebook component for single items}
\usage{
codebook_component_single_item(item, item_name, indent = "##")
}
\arguments{
\item{item}{an item with attributes set}

\item{item_name}{the item name}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
Codebook component for single items
}
\examples{
# will generate figure in a temporary directory
old_base_dir <- knitr::opts_knit$get("base.dir")
knitr::opts_knit$set(base.dir = tempdir())
on.exit(knitr::opts_knit$set(base.dir = old_base_dir))
data("bfi")
codebook_component_single_item(bfi$BFIK_open_1, "BFIK_open_1")
}
