% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_data_info}
\alias{codebook_data_info}
\title{Codebook data info}
\usage{
codebook_data_info(results, indent = "##")
}
\arguments{
\item{results}{a data frame which has the following columns: session, created, modified, expired, ended}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h2}
}
\description{
A readout of the metadata for this dataset, with some defaults set
}
\examples{
# will generate figures in a figure/ subdirectory
data("bfi")
metadata(bfi)$name <- "MOCK Big Five Inventory dataset (German metadata demo)"
metadata(bfi)$description <- "a small mock Big Five Inventory dataset"
metadata(bfi)$citation <- "doi:10.5281/zenodo.1326520"
metadata(bfi)$url <-
   "https://rubenarslan.github.io/codebook/articles/codebook.html"
codebook_data_info(bfi)
}
