% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook_table}
\alias{codebook_table}
\title{Codebook metadata table}
\usage{
codebook_table(results)
}
\arguments{
\item{results}{a data frame, ideally with attributes set on variables}
}
\description{
will generate a table combining metadata from variable attributes
with data summaries generated using \code{\link[skimr:skim_to_wide]{skimr::skim_to_wide()}}
}
\examples{
data("bfi")
codebook_table(bfi)
}
