% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{compact_codebook}
\alias{compact_codebook}
\title{Compact Codebook}
\usage{
compact_codebook(results)
}
\arguments{
\item{results}{the data frame}
}
\description{
Generate only the tabular codebook and the machine-readable JSON-LD metadata.
}
\examples{
# will generate figures in a figure/ subdirectory
old_base_dir <- knitr::opts_knit$get("base.dir")
knitr::opts_knit$set(base.dir = tempdir())
on.exit(knitr::opts_knit$set(base.dir = old_base_dir))
data("bfi")
bfi <- bfi[, c("BFIK_open_1", "BFIK_open_2")]
compact_codebook(bfi)
}
