% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{knit_print.alpha}
\alias{knit_print.alpha}
\title{Pretty-print a Cronbach's alpha object}
\usage{
knit_print.alpha(x, indent = "######", ...)
}
\arguments{
\item{x}{a psych alpha object}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h5}

\item{...}{ignored}
}
\description{
Turn a \code{\link[psych:alpha]{psych::alpha()}} object into HTML tables.
}
\examples{
example("alpha", "psych")
knitr::knit_print(a4)

}
