% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psych_print.R
\name{knit_print.multilevel}
\alias{knit_print.multilevel}
\title{Print a \code{\link[psych:multilevel.reliability]{psych::multilevel.reliability()}} object for knitr}
\usage{
\method{knit_print}{multilevel}(x, indent = "#####", ...)
}
\arguments{
\item{x}{a psych alpha object}

\item{indent}{add # to this to make the headings in the components lower-level. defaults to beginning at h5}

\item{...}{ignored}
}
\description{
Just prints the normal output of \code{\link[psych:multilevel.reliability]{psych::multilevel.reliability()}}.
}
\examples{
example("mlr", "psych")
knitr::knit_print(mg)

}
