% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown_helpers.R
\name{load_data_and_render_codebook}
\alias{load_data_and_render_codebook}
\title{Render codebook based on file}
\usage{
load_data_and_render_codebook(file, text, remove_file = FALSE, ...)
}
\arguments{
\item{file}{path to a file to make codebook from (sav, rds, dta, por, xpt, csv, csv2, tsv, etc.)}

\item{text}{codebook template}

\item{remove_file}{whether to remove file after rendering}

\item{...}{all other arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}}
}
\description{
Submit a data file and an rmarkdown template as a file to generate a codebook.
Used chiefly in the webapp.
}
