% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{metadata_jsonld}
\alias{metadata_jsonld}
\title{Metadata as JSON-LD}
\usage{
metadata_jsonld(results)
}
\arguments{
\item{results}{a data frame, ideally with attributes set on variables}
}
\description{
Echo a list of a metadata, generated using \code{\link[=metadata_list]{metadata_list()}} as JSON-LD in a
script tag.
}
\examples{
data("bfi")
metadata_jsonld(bfi)
}
