% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{metadata_list}
\alias{metadata_list}
\title{Metadata from dataframe}
\usage{
metadata_list(results)
}
\arguments{
\item{results}{a data frame}
}
\description{
Returns a list containing variable metadata (attributes) and data summaries.
}
