% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metadata.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_factor}
\alias{var_label}
\alias{var_label<-}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{haven}{\code{\link[haven]{as_factor}}}

  \item{labelled}{\code{\link[labelled]{var_label}}, \code{\link[labelled]{var_label<-}}}
}}

