% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.labelled}
\alias{summary.labelled}
\title{Summary function for labelled vector}
\usage{
\method{summary}{labelled}(object, ...)
}
\arguments{
\item{object}{a labelled vector}

\item{...}{passed to summary.factor}
}
\description{
Summary function for labelled vector
}
\examples{
example("labelled", "haven")
summary(x)

}
