% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summary.labelled_spss}
\alias{summary.labelled_spss}
\title{Summary function for labelled_spss vector}
\usage{
\method{summary}{labelled_spss}(object, ...)
}
\arguments{
\item{object}{a labelled_spss vector}

\item{...}{passed to summary.factor}
}
\description{
Summary function for labelled_spss vector
}
\examples{
example("labelled", "haven")
summary(x)

}
