% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_attributes.R
\name{zap_label}
\alias{zap_label}
\title{Zap variable label}
\usage{
zap_label(x)
}
\arguments{
\item{x}{the data frame or variable}
}
\description{
Modelled on \code{\link[haven:zap_labels]{haven::zap_labels()}}, zaps variable labels (not value labels).
}
